requirements.txt
asgiref==3.9.1
Django==5.2.6
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
